pyproject.toml
[project]
name = "jarvisx"
version = "0.1.0"
description = "JarvisX"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.22",
    "aiohttp>=3.9.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "asyncclick>=8.3.0.7",
    "click>=8.3.0",
    "google-adk>=1.22.0",
    "litellm>=1.80.13",
    "mcp>=1.18.0",
    "openai>=1.0.0",
    "psutil>=6.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.0",
    "greenlet>=3.0.0",
    "starlette>=0.27.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "httpx[http2]>=0.25.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.0.0",
    "pgvector>=0.3.0",
    "tiktoken>=0.7.0",
    "pypdf>=4.0.0",
    "python-multipart>=0.0.9",
    "apscheduler>=3.10.0",
    "langfuse>=2.0.0,<3.0.0",
    "reportlab>=4.4.9",
    "python-docx>=1.1.0",
    "pillow>=10.0.0",
    "authlib>=1.3.0",
    "python3-saml>=1.16.0",
    "cryptography>=42.0.0",
    "redis>=5.0.0",
]
